+++ tests/src/FunctionalJavascript/Integration/AutosaveFormTest.php
@@ -5,6 +5,7 @@ namespace Drupal\Tests\thunder\FunctionalJavascript\Integration;
 use Drupal\Tests\thunder\FunctionalJavascript\ThunderFormFieldTestTrait;
 use Drupal\Tests\thunder\FunctionalJavascript\ThunderJavascriptTestBase;
 use Drupal\Tests\thunder\FunctionalJavascript\ThunderParagraphsTestTrait;
+use Drupal\Tests\thunder\FunctionalJavascript\ThunderArticleTestTrait;
 
 /**
  * Tests the autosave support for nodes in Thunder.
@@ -15,6 +16,7 @@ class AutosaveFormTest extends ThunderJavascriptTestBase {
 
   use ThunderFormFieldTestTrait;
   use ThunderParagraphsTestTrait;
+  use ThunderArticleTestTrait;
 
   /**
    * {@inheritdoc}
