+++ tests/src/FunctionalJavascript/ThunderArticleTestTrait.php
@@ -11,7 +11,6 @@ trait ThunderArticleTestTrait {
 
   use ThunderFormFieldTestTrait;
   use ThunderJavascriptTrait;
-  use ThunderMediaTestTrait;
 
   /**
    * Pre-fill defined article fields for new article.
@@ -20,10 +19,6 @@ trait ThunderArticleTestTrait {
    *   Field values for new article.
    */
   public function articleFillNew(array $fieldValues) {
-    if (!$this instanceof ThunderJavascriptTestBase) {
-      throw new \RuntimeException('Trait is not used in correct context.');
-    }
-
     $this->drupalGet('node/add/article');
     $this->assertWaitOnAjaxRequest();
 
@@ -36,7 +31,38 @@ trait ThunderArticleTestTrait {
       }
       $this->setFieldValues($this->getSession()->getPage(), $fieldValues);
     }
+  }
 
+  /**
+   * Expand all tabs on page.
+   *
+   * It goes up to level 3 by default.
+   *
+   * @param int $maxLevel
+   *   Max depth of nested collapsed tabs.
+   */
+  public function expandAllTabs($maxLevel = 3) {
+    $jsScript = 'jQuery(\'details.js-form-wrapper.form-wrapper:not([open]) > summary\').click().length';
+
+    $numOfOpen = $this->getSession()->evaluateScript($jsScript);
+    $this->assertWaitOnAjaxRequest();
+
+    for ($i = 0; $i < $maxLevel && $numOfOpen > 0; $i++) {
+      $numOfOpen = $this->getSession()->evaluateScript($jsScript);
+      $this->assertWaitOnAjaxRequest();
+    }
+  }
+
+  /**
+   * Set moderation state.
+   *
+   * @param string $state
+   *   State id.
+   */
+  protected function setModerationState($state) {
+    $this->getSession()
+      ->getDriver()
+      ->selectOption('//*[@id="edit-moderation-state-0"]', $state);
   }
 
 }
