+++ tests/src/FunctionalJavascript/ThunderEntityBrowserTestTrait.php
@@ -3,7 +3,6 @@
 namespace Drupal\Tests\thunder\FunctionalJavascript;
 
 use Behat\Mink\Element\DocumentElement;
-use Behat\Mink\Element\NodeElement;
 
 /**
  * Trait with support for handling Entity Browser actions.
@@ -11,6 +10,8 @@ use Behat\Mink\Element\NodeElement;
  * @package Drupal\Tests\thunder\FunctionalJavascript
  */
 trait ThunderEntityBrowserTestTrait {
+
+  use ThunderJavascriptTrait;
 
   /**
    * Open modal entity browser and switch into iframe from it.
@@ -23,7 +24,8 @@ trait ThunderEntityBrowserTestTrait {
    *   Entity browser name.
    */
   public function openEntityBrowser(DocumentElement $page, $drupalSelector, $entityBrowser) {
-    $this->clickButtonDrupalSelector($page, $drupalSelector);
+    $this->clickButtonDrupalSelector($page, $drupalSelector . '-entity-browser-entity-browser-open-modal');
+    $this->assertWaitOnAjaxRequest();
 
     $this->getSession()
       ->switchToIFrame('entity_browser_iframe_' . $entityBrowser);
@@ -39,9 +41,22 @@ trait ThunderEntityBrowserTestTrait {
    *
    * @param \Behat\Mink\Element\DocumentElement $page
    *   Current active page.
+   * @param string $entityBrowser
+   *   Entity browser name.
    */
-  public function submitEntityBrowser(DocumentElement $page) {
-    $this->clickButtonDrupalSelector($page, 'edit-use-selected', FALSE);
+  public function submitEntityBrowser(DocumentElement $page, $entityBrowser) {
+    if ($entityBrowser == 'multiple_image_browser') {
+      $this->getSession()->wait(200);
+      $this->assertWaitOnAjaxRequest();
+
+      $page->pressButton('Use selected');
+    }
+    elseif ($entityBrowser == 'image_browser') {
+      $page->pressButton('Select image');
+    }
+    elseif ($entityBrowser == 'video_browser') {
+      $page->pressButton('Select video');
+    }
 
     $this->getSession()->switchToIFrame();
     $this->assertWaitOnAjaxRequest();
@@ -117,41 +132,6 @@ trait ThunderEntityBrowserTestTrait {
     // command for auto selection are executed.
     $this->getSession()->wait(200);
     $this->assertWaitOnAjaxRequest();
-  }
-
-  /**
-   * Drag element in document with defined offset position.
-   *
-   * @param \Behat\Mink\Element\NodeElement $element
-   *   Element that will be dragged.
-   * @param int $offsetX
-   *   Vertical offset for element drag in pixels.
-   * @param int $offsetY
-   *   Horizontal offset for element drag in pixels.
-   */
-  protected function dragDropElement(NodeElement $element, $offsetX, $offsetY) {
-    $this->assertWaitOnAjaxRequest();
-
-    $elemXpath = $element->getXpath();
-
-    $jsCode = "var fireMouseEvent = function (type, element, x, y) {"
-      . "  var event = document.createEvent('MouseEvents');"
-      . "  event.initMouseEvent(type, true, (type !== 'mousemove'), window, 0, 0, 0, x, y, false, false, false, false, 0, element);"
-      . "  element.dispatchEvent(event); };";
-
-    // XPath provided by getXpath uses single quote (') to encapsulate strings,
-    // that's why xpath has to be quited with double quites in javascript code.
-    $jsCode .= "(function() {" .
-      "  var dragElement = document.evaluate(\"{$elemXpath}\", document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null).singleNodeValue;" .
-      "  var pos = dragElement.getBoundingClientRect();" .
-      "  var centerX = Math.floor((pos.left + pos.right) / 2);" .
-      "  var centerY = Math.floor((pos.top + pos.bottom) / 2);" .
-      "  fireMouseEvent('mousedown', dragElement, centerX, centerY);" .
-      "  fireMouseEvent('mousemove', document, centerX + {$offsetX}, centerY + {$offsetY});" .
-      "  fireMouseEvent('mouseup', dragElement, centerX + {$offsetX}, centerY + {$offsetY});" .
-      "})();";
-
-    $this->getSession()->executeScript($jsCode);
   }
 
 }
