+++ tests/src/FunctionalJavascript/ThunderFormFieldTestTrait.php
@@ -11,6 +11,8 @@ use Behat\Mink\Element\DocumentElement;
  */
 trait ThunderFormFieldTestTrait {
 
+  use ThunderJavascriptTrait;
+
   /**
    * Set value for group of checkboxes.
    *
@@ -112,6 +114,24 @@ trait ThunderFormFieldTestTrait {
     foreach ($fieldValues as $fieldName => $value) {
       $this->setFieldValue($page, $fieldName, $value);
     }
+  }
+
+  /**
+   * Set value directly to field value, without formatting applied.
+   *
+   * @param string $fieldName
+   *   Field name.
+   * @param string $rawValue
+   *   Raw value for field.
+   */
+  public function setRawFieldValue($fieldName, $rawValue) {
+    // Set date over jQuery, because browser drivers handle input value
+    // differently. fe. (Firefox will set it as "value" for field, but Chrome
+    // will use it as text for that input field, and in that case final value
+    // depends on format used for input field. That's why it's better to set it
+    // directly to value, independently from format used.
+    $this->getSession()
+      ->executeScript("jQuery('[name=\"{$fieldName}\"]').val('{$rawValue}')");
   }
 
 }
