+++ tests/src/FunctionalJavascript/ThunderMediaTestTrait.php
@@ -9,6 +9,8 @@ namespace Drupal\Tests\thunder\FunctionalJavascript;
  */
 trait ThunderMediaTestTrait {
 
+  use ThunderEntityBrowserTestTrait;
+
   /**
    * Select Medias for field.
    *
@@ -23,42 +25,16 @@ trait ThunderMediaTestTrait {
     $page = $this->getSession()->getPage();
     $driver = $this->getSession()->getDriver();
 
-    $this->assertWaitOnAjaxRequest();
-
-    $buttonName = $fieldName . '_entity_browser_entity_browser';
-    $this->scrollElementInView("[name=\"{$buttonName}\"]");
-    $page->pressButton($buttonName);
-
-    $this->assertWaitOnAjaxRequest();
-
-    $this->getSession()
-      ->switchToIFrame('entity_browser_iframe_' . $entityBrowser);
-    $this->assertWaitOnAjaxRequest();
+    $this->openEntityBrowser($page, $this->fieldNameToDrupalSelector($fieldName), $entityBrowser);
 
     foreach ($medias as $media) {
       $driver->click("//div[contains(@class, 'views-row') and .//*[@name='entity_browser_select[$media]']]");
     }
     $this->assertWaitOnAjaxRequest();
 
-    $element = 'img';
-    if ($entityBrowser == 'multiple_image_browser') {
-      $this->getSession()->wait(200);
-      $this->assertWaitOnAjaxRequest();
+    $this->submitEntityBrowser($page, $entityBrowser);
 
-      $page->pressButton('Use selected');
-    }
-    elseif ($entityBrowser == 'image_browser') {
-      $page->pressButton('Select image');
-    }
-    elseif ($entityBrowser == 'video_browser') {
-      $page->pressButton('Select video');
-      $element = 'iframe';
-    }
-
-    $this->getSession()->switchToIFrame();
-    $this->assertWaitOnAjaxRequest();
-
-    $this->waitUntilVisible('div[data-drupal-selector="edit-' . str_replace('_', '-', $fieldName) . '-wrapper"] ' . $element);
+    $this->assertSession()->waitForElementVisible('css', 'div[data-drupal-selector="edit-' . str_replace('_', '-', $fieldName) . '-wrapper"]');
   }
 
   /**
